Move "freeSpaceIfNearQuota" out of background.js
